refactor: move user to module
