+++ backend/src/routes/userRouter.ts
@@ -1,8 +1,7 @@
 import express from 'express';
-import { UserController } from '../controllers/UserController.js';
-import { UserServices } from '../services/UserServices.js';
-import { UserRepository } from '../repositories/UserRepository.js';
-import { UserValidator } from '../middlewares/userValidator.js';
+import { UserController } from '../users/controllers/UserController.js';
+import { UserServices } from '../users/services/UserServices.js';
+import { UserRepository } from '../users/repositories/UserRepository.js';
 
 const userRepository = new UserRepository();
 const userService = new UserServices(userRepository);
@@ -15,8 +14,7 @@ router.post('/', userController.create.bind(userController));
 
 // TODO: Añadir middleware de autenticación para rutas protegidas
 router.get('/:id', userController.findById.bind(userController));
-router.get('/email/:email', userController.findByEmail.bind(userController));
 router.put('/:id', userController.update.bind(userController));
 router.delete('/:id', userController.delete.bind(userController));
 
-export default router; 
+export default router;
