+++ backend/src/users/controllers/IUserController.ts
@@ -1,0 +1,8 @@
+import { Request, Response, NextFunction } from 'express';
+
+export interface IUserController {
+  findById(req: Request, res: Response, next: NextFunction): Promise<void>;
+  create(req: Request, res: Response, next: NextFunction): Promise<void>;
+  update(req: Request, res: Response, next: NextFunction): Promise<void>;
+  delete(req: Request, res: Response, next: NextFunction): Promise<void>;
+}
