+++ backend/src/users/exceptions/EmailAlreadyExists.exception.ts
@@ -1,0 +1,11 @@
+import { BaseException } from '../../shared/exceptions/BaseException';
+import { ErrorCode } from '../../shared/exceptions/ErrorCode.enum';
+
+export class EmailAlreadyExistsException extends BaseException {
+  constructor(email: string) {
+    super(
+      `User with email ${email} already exists`,
+      ErrorCode.EMAIL_ALREADY_EXISTS,
+    );
+  }
+}
