+++ backend/src/users/exceptions/UserNotFoundException.ts
@@ -1,0 +1,8 @@
+import { BaseException } from '../../../shared/exceptions/BaseException.js';
+import { ErrorCode } from '../../../shared/exceptions/ErrorCode.enum.js';
+
+export class UserNotFoundException extends BaseException {
+  constructor(userId: string) {
+    super(`User with id ${userId} not found`, ErrorCode.USER_NOT_FOUND);
+  }
+}
