+++ backend/src/users/models/dtos/UserDTO.ts
@@ -5,7 +5,6 @@ export interface CreateUserDTO {
 }
 
 export interface UpdateUserDTO {
-  id: string;
   username?: string;
   email?: string;
 }
