+++ backend/src/users/repositories/IUserRepository.ts
@@ -1,4 +1,4 @@
-import { User } from "../../models/entities/User.js";
+import { User } from "../models/entities/User.js";
 
 export interface IUserRepository {
   findById(id: string): Promise<User | undefined>;
